+++ internal/kibana/synthetics/schema.go
@@ -1,0 +1,52 @@
+package synthetics
+
+import (
+	"github.com/disaster37/go-kibana-rest/v8/kbapi"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+const (
+	MetadataPrefix = "_kibana_synthetics_"
+)
+
+func GeoConfigSchema() schema.Attribute {
+	return schema.SingleNestedAttribute{
+		Optional:    true,
+		Description: "Geographic coordinates (WGS84) for the location",
+		Attributes: map[string]schema.Attribute{
+			"lat": schema.Float64Attribute{
+				Optional:            false,
+				Required:            true,
+				MarkdownDescription: "The latitude of the location.",
+			},
+			"lon": schema.Float64Attribute{
+				Optional:            false,
+				Required:            true,
+				MarkdownDescription: "The longitude of the location.",
+			},
+		},
+	}
+}
+
+type TFGeoConfigV0 struct {
+	Lat types.Float64 `tfsdk:"lat"`
+	Lon types.Float64 `tfsdk:"lon"`
+}
+
+func (m *TFGeoConfigV0) ToSyntheticGeoConfig() *kbapi.SyntheticGeoConfig {
+	return &kbapi.SyntheticGeoConfig{
+		Lat: m.Lat.ValueFloat64(),
+		Lon: m.Lon.ValueFloat64(),
+	}
+}
+
+func FromSyntheticGeoConfig(v *kbapi.SyntheticGeoConfig) *TFGeoConfigV0 {
+	if v == nil {
+		return nil
+	}
+	return &TFGeoConfigV0{
+		Lat: types.Float64Value(v.Lat),
+		Lon: types.Float64Value(v.Lon),
+	}
+}