+++ libs/go-kibana-rest/kbapi/api.kibana_synthetics_test.go
@@ -119,6 +119,7 @@ func (s *KBAPITestSuite) TestKibanaSyntheticsMonitorAPI() {
 							Namespace:      space,
 							Params: map[string]interface{}{
 								"param1": "some-params",
+								"my_url": "http://localhost:8080",
 							},
 							RetestOnFailure: f,
 						},
@@ -179,6 +180,7 @@ func (s *KBAPITestSuite) TestKibanaSyntheticsMonitorAPI() {
 					monitor, err := syntheticsAPI.Monitor.Add(config, fields, space)
 					assert.NoError(s.T(), err)
 					assert.NotNil(s.T(), monitor)
+					monitor.Params = nil //kibana API doesn't return params for GET request
 
 					get, err := syntheticsAPI.Monitor.Get(monitor.Id, space)
 					assert.NoError(s.T(), err)
@@ -190,6 +192,8 @@ func (s *KBAPITestSuite) TestKibanaSyntheticsMonitorAPI() {
 
 					update, err := syntheticsAPI.Monitor.Update(monitor.Id, tc.update.config, tc.update.fields, space)
 					assert.NoError(s.T(), err)
+					assert.NotNil(s.T(), update)
+					update.Params = nil //kibana API doesn't return params for GET request
 
 					get, err = syntheticsAPI.Monitor.Get(monitor.ConfigId, space)
 					assert.NoError(s.T(), err)
@@ -252,5 +256,24 @@ func (s *KBAPITestSuite) TestKibanaSyntheticsPrivateLocationAPI() {
 				assert.Error(s.T(), err)
 			})
 		})
+	}
+}
+
+func (s *KBAPITestSuite) TestKibanaSyntheticsPrivateLocationNotFound() {
+	for _, n := range namespaces {
+		testUuid := uuid.New().String()
+		space := n
+		pAPI := s.API.KibanaSynthetics.PrivateLocation
+
+		ids := []string{"", "not-found", testUuid}
+
+		for _, id := range ids {
+			s.Run(fmt.Sprintf("TestKibanaSyntheticsPrivateLocationNotFound - %s - %s", n, id), func() {
+				_, err := pAPI.Get(id, space)
+				assert.Error(s.T(), err)
+				assert.IsType(s.T(), APIError{}, err)
+				assert.Equal(s.T(), 404, err.(APIError).Code)
+			})
+		}
 	}
 }