+++ provider/plugin_framework.go
@@ -7,11 +7,17 @@ import (
 	"github.com/elastic/terraform-provider-elasticstack/internal/clients/config"
 	"github.com/elastic/terraform-provider-elasticstack/internal/kibana/data_view"
 	"github.com/elastic/terraform-provider-elasticstack/internal/kibana/import_saved_objects"
+	"github.com/elastic/terraform-provider-elasticstack/internal/kibana/synthetics/private_location"
 	"github.com/elastic/terraform-provider-elasticstack/internal/schema"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	fwprovider "github.com/hashicorp/terraform-plugin-framework/provider"
 	fwschema "github.com/hashicorp/terraform-plugin-framework/provider/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
+)
+
+// Ensure the implementation satisfies the expected interfaces.
+var (
+	_ fwprovider.Provider = &Provider{}
 )
 
 type Provider struct {
@@ -41,14 +47,14 @@ func (p *Provider) Schema(ctx context.Context, req fwprovider.SchemaRequest, res
 }
 
 func (p *Provider) Configure(ctx context.Context, req fwprovider.ConfigureRequest, res *fwprovider.ConfigureResponse) {
-	var config config.ProviderConfiguration
+	var cfg config.ProviderConfiguration
 
-	res.Diagnostics.Append(req.Config.Get(ctx, &config)...)
+	res.Diagnostics.Append(req.Config.Get(ctx, &cfg)...)
 	if res.Diagnostics.HasError() {
 		return
 	}
 
-	client, diags := clients.NewApiClientFromFramework(ctx, config, p.version)
+	client, diags := clients.NewApiClientFromFramework(ctx, cfg, p.version)
 	res.Diagnostics.Append(diags...)
 	if res.Diagnostics.HasError() {
 		return
@@ -66,5 +72,6 @@ func (p *Provider) Resources(ctx context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		func() resource.Resource { return &import_saved_objects.Resource{} },
 		func() resource.Resource { return &data_view.Resource{} },
+		func() resource.Resource { return &private_location.Resource{} },
 	}
 }